Upload to master

+++ CCP/plugin/network_service/fermat-ccp-plugin-network-service-crypto-addresses-bitdubai/src/main/java/com/bitdubai/fermat_ccp_plugin/layer/network_service/crypto_addresses/developer/bitdubai/version_1/CryptoAddressesNetworkServicePluginRoot.java
@@ -61,6 +61,7 @@ import com.bitdubai.fermat_ccp_plugin.layer.network_service.crypto_addresses.dev
 import com.bitdubai.fermat_ccp_plugin.layer.network_service.crypto_addresses.developer.bitdubai.version_1.communication.structure.CommunicationRegistrationProcessNetworkServiceAgent;
 import com.bitdubai.fermat_ccp_plugin.layer.network_service.crypto_addresses.developer.bitdubai.version_1.database.CryptoAddressesNetworkServiceDao;
 import com.bitdubai.fermat_ccp_plugin.layer.network_service.crypto_addresses.developer.bitdubai.version_1.database.CryptoAddressesNetworkServiceDeveloperDatabaseFactory;
+import com.bitdubai.fermat_ccp_plugin.layer.network_service.crypto_addresses.developer.bitdubai.version_1.exceptions.CantChangeProtocolStateException;
 import com.bitdubai.fermat_ccp_plugin.layer.network_service.crypto_addresses.developer.bitdubai.version_1.exceptions.CantCreateRequestException;
 import com.bitdubai.fermat_ccp_plugin.layer.network_service.crypto_addresses.developer.bitdubai.version_1.exceptions.CantHandleNewMessagesException;
 import com.bitdubai.fermat_ccp_plugin.layer.network_service.crypto_addresses.developer.bitdubai.version_1.exceptions.CantInitializeCryptoAddressesNetworkServiceDatabaseException;
@@ -90,6 +91,8 @@ import com.google.gson.Gson;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Timer;
+import java.util.TimerTask;
 import java.util.UUID;
 import java.util.concurrent.CopyOnWriteArrayList;
 
@@ -245,6 +248,21 @@ public class CryptoAddressesNetworkServicePluginRoot extends AbstractNetworkServ
             }
 
             remoteNetworkServicesRegisteredList = new CopyOnWriteArrayList<>();
+
+            // change message state to process again first time
+            reprocessWaitingMessage();
+
+            //declare a schedule to process waiting request message
+            Timer timer = new Timer();
+
+            timer.schedule(new TimerTask() {
+                @Override
+                public void run() {
+                    // change message state to process again
+                    reprocessWaitingMessage();
+                }
+            }, 2*3600*1000);
+
 
             /*
              * Its all ok, set the new status
@@ -475,7 +493,8 @@ public class CryptoAddressesNetworkServicePluginRoot extends AbstractNetworkServ
                     action,
                     dealer,
                     blockchainNetworkType,
-                    1
+                    1,
+                    System.currentTimeMillis()
             );
 
             System.out.println("********* Crypto Addresses: Successful Address Exchange Request creation. ");
@@ -821,7 +840,7 @@ public class CryptoAddressesNetworkServicePluginRoot extends AbstractNetworkServ
         cryptoAddressesExecutorAgent.connectionFailure(remoteParticipant.getIdentityPublicKey());
 
         //I check my time trying to send the message
-        //checkFailedDeliveryTime(remoteParticipant.getIdentityPublicKey());
+        checkFailedDeliveryTime(remoteParticipant.getIdentityPublicKey());
 
     }
 
@@ -857,6 +876,21 @@ public class CryptoAddressesNetworkServicePluginRoot extends AbstractNetworkServ
 
             if(vpnConnectionCloseNotificationEvent.getNetworkServiceApplicant() == getNetworkServiceType()){
 
+                try {
+
+                    List<CryptoAddressRequest> cryptoAddressRequestList = cryptoAddressesNetworkServiceDao.listPendingRequestsByProtocolState(ProtocolState.WAITING_RESPONSE);
+
+                    for(CryptoAddressRequest record : cryptoAddressRequestList) {
+
+                        cryptoAddressesNetworkServiceDao.changeProtocolState(record.getRequestId(),ProtocolState.PROCESSING_SEND);
+                    }
+                }
+                catch(CantListPendingCryptoAddressRequestsException | CantChangeProtocolStateException |PendingRequestNotFoundException e)
+                {
+                    System.out.print("EXCEPCION REPROCESANDO WAIT MESSAGE");
+                    e.printStackTrace();
+                }
+
                 if(communicationNetworkServiceConnectionManager != null) {
                     communicationNetworkServiceConnectionManager.closeConnection(vpnConnectionCloseNotificationEvent.getRemoteParticipant().getIdentityPublicKey());
                 }
@@ -875,6 +909,22 @@ public class CryptoAddressesNetworkServicePluginRoot extends AbstractNetworkServ
 
         if(fermatEvent instanceof ClientConnectionCloseNotificationEvent){
             System.out.println("CLOSSING ALL CONNECTIONS IN CRYPTO ADDRESSES ");
+
+            try {
+
+                List<CryptoAddressRequest> cryptoAddressRequestList = cryptoAddressesNetworkServiceDao.listPendingRequestsByProtocolState(ProtocolState.WAITING_RESPONSE);
+
+                for(CryptoAddressRequest record : cryptoAddressRequestList) {
+
+                    cryptoAddressesNetworkServiceDao.changeProtocolState(record.getRequestId(),ProtocolState.PROCESSING_SEND);
+                }
+            }
+            catch(CantListPendingCryptoAddressRequestsException | CantChangeProtocolStateException |PendingRequestNotFoundException e)
+            {
+                System.out.print("EXCEPCION REPROCESANDO WAIT MESSAGE");
+                e.printStackTrace();
+            }
+
             this.register = false;
             if(communicationNetworkServiceConnectionManager != null) {
                 communicationNetworkServiceConnectionManager.closeAllConnection();
@@ -931,6 +981,11 @@ public class CryptoAddressesNetworkServicePluginRoot extends AbstractNetworkServ
         }
     }
 
+    @Override
+    public void handleNewSentMessageNotificationEvent(FermatMessage data) {
+
+    }
+
     /**
      * I indicate to the Agent the action that it must take:
      * - Protocol State: PROCESSING_RECEIVE.
@@ -957,7 +1012,8 @@ public class CryptoAddressesNetworkServicePluginRoot extends AbstractNetworkServ
                     action,
                     requestMessage.getCryptoAddressDealer(),
                     requestMessage.getBlockchainNetworkType(),
-                    1
+                    1,
+                    System.currentTimeMillis()
             );
 
         } catch(CantCreateRequestException e) {
@@ -1041,7 +1097,7 @@ public class CryptoAddressesNetworkServicePluginRoot extends AbstractNetworkServ
 
                 if(!record.getState().getCode().equals(ProtocolState.WAITING_RESPONSE.getCode()))
                 {
-                    if(record.getSentNumber() > 5 )
+                    if(record.getSentNumber() > 10)
                     {
                          //update state and process again later
                         cryptoAddressesNetworkServiceDao.changeProtocolState(record.getRequestId(),ProtocolState.WAITING_RESPONSE);
@@ -1053,15 +1109,21 @@ public class CryptoAddressesNetworkServicePluginRoot extends AbstractNetworkServ
                 }
                 else
                 {
-                    //I verify the number of days I'm around trying to send if it exceeds seven I delete record
+                    //I verify the number of days I'm around trying to send if it exceeds three days I delete record
 
-                    //long sentDate = record.get();
-                    //long currentTime = System.currentTimeMillis();
+                    long sentDate = record.getSentDate();
+                    long currentTime = System.currentTimeMillis();
+                    long dif = currentTime - sentDate;
 
-                    //long dif = currentTime - sentDate;
+                    double dias = Math.floor(dif / (1000 * 60 * 60 * 24));
 
-                    //if(dif > 604800000)
+                    if((int) dias > 3)
+                    {
+                        //notify the user does not exist to intra user actor plugin
+
                         cryptoAddressesNetworkServiceDao.delete(record.getRequestId());
+                    }
+
                 }
 
             }
@@ -1074,5 +1136,24 @@ public class CryptoAddressesNetworkServicePluginRoot extends AbstractNetworkServ
             e.printStackTrace();
         }
 
+    }
+
+
+    private void reprocessWaitingMessage()
+    {
+        try {
+
+            List<CryptoAddressRequest> cryptoAddressRequestList = cryptoAddressesNetworkServiceDao.listPendingRequestsByProtocolState(ProtocolState.WAITING_RESPONSE);
+
+            for(CryptoAddressRequest record : cryptoAddressRequestList) {
+
+                cryptoAddressesNetworkServiceDao.changeProtocolState(record.getRequestId(),ProtocolState.PROCESSING_SEND);
+            }
+        }
+        catch(CantListPendingCryptoAddressRequestsException | CantChangeProtocolStateException |PendingRequestNotFoundException e)
+        {
+            System.out.print("EXCEPCION REPROCESANDO WAIT MESSAGE");
+            e.printStackTrace();
+        }
     }
 }
